version 1

+++ pages/api/admin/dev.ts
@@ -21,8 +21,8 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse<
 
     }
     if (req.method == "POST") {
-
         try {
+            if (!process.env.IS_ADMIN ) throw new Error('Unauthorized')
             const { name, summary, logo, location, country } = req.body; 
             const msg = req.body.id ? "Update successful" : "Creation successful";
             const id = req.body.id || new mongoose.mongo.ObjectId().toString(); ;
@@ -47,6 +47,7 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse<
     }
     if (req.method == "DELETE") {
         try {
+            if (!process.env.IS_ADMIN ) throw new Error('Unauthorized')        
             const {id} = req.body;
             const result = await Developers.findByIdAndDelete(id);
             if (result == null) throw new Error("Item not found")
